Green Hell: Fix depth buffer & enable stereo compute shaders
One of the depth buffers was mono, leading to what appeared to be a halo
issue on the main menu. Force the depth buffer to stereo in the template
unity.ini to fix this.

This is the first Unity game I've seen using compute shaders, so enable
stereo compute shaders to fix some of the lighting issues on the main
menu.

The fog shader is temporarily disabled to make the main menu visible -
this will be reverted in the following commit.