Green Hell: Revert & re-enable fog
This removes the autofix from the fog shader and re-enables it. It seems
this shader requires either or both of the depth buffer and stereo
compute shader fixes to render correctly, hence it was easier to disable
it until those were resolved.

I'm not yet positive if this may require some tweaking to look perfect -
I'll examine it closer once the god rays are resolved.