Green Hell: Fix Directional lighting
This is a different pattern to usual, that does not directly calculate
the linear depth using _ZBufferParams, but does use the inverse
projection matrix to calculate view-space before moving on to world
space. Add the adjustment at that point instead since we have the
required linear depth and inverse projection matrix.

May need to add this variant to the autofix as there does appear to be
126 variants of this shader.