Green Hell: Add a crosshair adjustment
This applies a crosshair style adjustment to the entire HUD whenever the
cursor is hidden. This is just a first pass, and we might do something
more sophisticated later.

This uses the depth buffer & inverse projection matrix copied from the
directional lighting shader. The shaders chosen by the template seem
to be used for reflection probes once a second making them unreliable.
This might not be final.

A debugging shader is added to check the z_to_w maths.