Green Hell: Tweak crosshair adjustment to work with AA on/off
A different lighting pixel shader is used with AA off, so the
ResourceDepth is not filled out. Using the resource from the template
works much better with AA off, but breaks when certain objects (e.g.
rocks) are highlighted as the game renders them to a separate depth
buffer.

Change the logic to use the directional lighting vertex shader hash to
grab the depth buffer, adding a check to ensure the depth buffer is
rectangular to avoid the reflection probes, and only grab the first in
the frame to avoid the secondary depth buffer used for highlighted
objects.

Also pass in the UnityPerCameraRare buffer from the template, as this
seems to be more reliable than UnityPerCamera for this game.