Green Hell: Enable software mouse
I think in this game we won't want the HUD adjustment jumping back to
screen depth whenever the mouse cursor is visible. Enable the software
mouse and apply the crosshair adjustment to it.

We might want something like center weighted average for the main UI
adjustment later, as the grass snags the crosshair depth too much.

Switches to the alternate alpha mode on the cursor shader.