Green Hell: Tweak Walkie-Talkie preset
Drop the keyboard preset as it doesn't play nice with the texture preset
and that will need some improvements in 3DMigoto to address. Instead,
add the hashes for the powered off walkie-talkie as well and use texture
matches for everything. Boost up the convergence preset a little since
this can be used while walking.