Green Hell: Adjust HUD based on center weighted average
Rather than using the crosshair depth for the entire HUD, which can get
snagged on a blade of grass, downscale and average the center of the
depth buffer and use it to select a more appropriate depth.

Currently the crosshair is also set to this depth, but we disable it
because it's useless - if it turns out to be useful later in the game we
can do something better.