Green Hell: Fix periodic ground flimmering
The ground uses a parallax shader in this game, but flimmers
periodically, in similar timing to the reflection probes we saw earlier.

Add render target size checks when grabbing any unity resources to avoid
grabbing invalid matrices and fix the issue.

Note that this will be clobbered if the autofix is re-run - we might
consider permanently adding this check to the autofix!