Green Hell: Use outline depth buffer when available
When an object outline is being drawn we want the HUD depth to line up
with the object. This uses a separate depth buffer, but the outline
shader is always run regardless of whether an object is actually
highlighted or not making it non-trivial to determine which depth buffer
to use. This downscales both depth buffers - the outline depth buffer is
downscaled from full coverage, while the other depth buffer is only
downscaled in the center. The HUD shader will check if the downscaled
outline buffer contains any data, and will use it if so, otherwise will
fall back to the downscaled regular depth buffer.