Pop backpack and crafting interface into 3D
The backpack inventory & crafting table are rendered to an unusual sized
render target that the driver was not steroising. Force these to stereo
so that the crafting interface is in 3D.

The depth is not ideal, but is workable. The HUD will need to detect
this mode and adjust to a more appropriate depth, provided the outline
shader is not currently in use.